Added BlogFragment and PostFragment

+++ app/src/main/java/io/github/mamedovilkin/finexetf/model/view/Asset.kt
@@ -6,7 +6,7 @@ data class Asset(
     val name: String,
     val originalName: String,
     val navPrice: Double,
-    val totalQuantity: Int,
+    val totalQuantity: Long,
     val totalPrice: Double,
     val totalNavPrice: Double,
 )
